Add dump file classes for all filetypes
