+++ src/fst.h
@@ -22,11 +22,5 @@
 typedef struct _FstFile FstFile;
 
 GwDumpFile *fst_main(char *fname, char *skip_start, char *skip_end);
-void import_fst_trace(FstFile *self, GwNode *np);
-void fst_set_fac_process_mask(FstFile *self, GwNode *np);
-void fst_import_masked(FstFile *self);
-
-void fst_file_close(FstFile *self);
-gchar *fst_file_get_subvar(FstFile *self, gint index);
 
 #endif
