+++ src/gw-ghw-file-private.h
@@ -1,0 +1,9 @@
+#pragma once
+
+struct _GwGhwFile
+{
+    GwDumpFile parent_instance;
+
+    GwHistEntFactory *hist_ent_factory;
+};
+
